requirements.txt
absl-py==1.4.0
cachetools==4.2.4
charset-normalizer==3.1.0
einops==0.4.0
fvcore==0.1.5.post20220119
gitdb==4.0.10
GitPython==3.1.31
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.53.0
idna==3.4
imageio==2.9.0
importlib-metadata==6.4.1
iopath==0.1.10
Markdown==3.4.3
MarkupSafe==2.1.2
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.5.5.62
Pillow==9.5.0
portalocker==2.7.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
rawpy==0.17.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
smmap==5.0.0
tabulate==0.9.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
timm==0.5.4
torch==1.10.1+cu111
torchaudio==0.10.1+cu111
torchvision==0.11.2+cu111
tqdm==4.62.3
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yacs==0.1.8
zipp==3.15.0
setuptools==58.0.4
